requirements.txt
boto3==1.9.197
botocore==1.12.197
Django==2.2.4
django-filter==2.2.0
django-guardian==2.0.0
django-split-settings==0.3.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
jmespath==0.9.4
pkg-resources==0.0.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
